Remove unused parameters from Parameters.cpp change email from martin.steinegger@campus.lmu.de -> martin.steinegger@mpibpc.mpg.de add automatic sens setup for clustering workflow remove some warnings change calculation for sens in prefilter to upper. limit - (sens * 8.75)

+++ src/prefiltering/Prefiltering.h
@@ -86,7 +86,7 @@ private:
     int kmerSize;
     const int kmerScore;
     bool spacedKmer;
-    const int sensitivity;
+    const float sensitivity;
     size_t maxResListLen;
     int alphabetSize;
     const float zscoreThr;
@@ -128,7 +128,7 @@ private:
 
     void mergeFiles(std::vector<std::pair<std::string, std::string>> splitFiles, int mode);
 
-    const int getKmerThreshold(const int sensitivity, const int score);
+    const int getKmerThreshold(const float sensitivity, const int score);
 };
 
 #endif
